Add a minimalistic parser combinators library
Parser combinators are a simple, efficient, flexible,
and reusable of parsing both textual and binary formats.
Most of the code in this commit is either inspired, or
imported from github.com/Jeffail/benthos blob language's
implementation (MIT License). It is was the best option
when it comes to combinators in Go, but was unfortunately
not available an independent library/module.